fix: get_event_from_transaction
