+++ src/contracts_caller/types.rs
@@ -25,12 +25,3 @@ impl TryInto<VerifyMerkle> for Event {
         })
     }
 }
-
-
-
-
-
-
-
-
-
