+++ src/contracts_caller/verify_fri/mod.rs
@@ -1,6 +1,8 @@
 pub mod compute_next_layer;
-pub mod verify_merkle;
+pub mod merkle_verifier;
 pub mod init_fri_group;
-pub mod verify_fri;
+pub mod fri_statement;
 pub mod register_fact_fri;
 pub mod types;
+pub mod sample_verify_fri_input;
+pub mod verify_fri;
