+++ src/contracts_caller/verify_fri/types/mod.rs
@@ -1,0 +1,4 @@
+pub mod compute_next_layer;
+pub mod fri_verify_input;
+pub mod init_fri_group;
+pub mod register_fact_verify_fri;
