+++ src/contracts_caller/verify_merkle/mod.rs
@@ -1,3 +1,5 @@
 pub mod merkle_statement;
 pub mod register_fact_merkle;
 pub mod types;
+pub mod sample_verify_merkle_input;
+pub mod verify_merkle;
