+++ src/contracts_caller/verify_merkle/types/mod.rs
@@ -1,0 +1,2 @@
+pub mod register_fact_verify_merkle;
+pub mod verify_merkle_input;
