+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod contracts_caller;
 pub mod config;
-pub mod verify_flow;
+pub mod flow_test;
 pub mod error;
